Rename `expectEventWithPropertiesFilter` function, add JSDoc and adjust a callback argument to make its purpose clearer.
Address: https://github.com/woocommerce/google-listings-and-ads/pull/2419#discussion_r1626502194
